fix: reconcile the merged spec tiers and green the gates
Follow-up to the main merge. Everything below was found by running the gates
on the merged tree — this branch's CI had never reached them, because
check:all failed at lint:mock-boundaries long before test:coverage.

Import rewiring
- `diff` moved to conformance/projectionDiff.ts on main (it was split out of
  semanticProjection to stay under the file-size limit). The spec gates and
  triage dump still imported it from the old module, so every spec example
  failed with "diff is not a function" — 3684 failures from one moved symbol.

Ledger reconciliation (the two-way ratchet doing its job)
- 133 declarations removed as STALE: main fixed the defects they recorded —
  the leading-`---` frontmatter trap (cm-43/47/77), bracket-escape growth
  (cm-194/512/549/550), the `about:blank` URL rewrites (cm-500/596/…), and
  reference flattening on images. Each removal is a recorded win.
- 5 declarations added: image and link references now SURVIVE a save, but a
  shortcut/collapsed reference is re-emitted in full form when its label and
  rendered alt differ. Rendered output and resolved destination are
  unchanged; only the spelling widens. Verdict `normalization`.

XSS oracle updated to the current policy
- specXss asserted every round-tripped URL is `isSafeUrl` or `about:blank` —
  the storage-rewrite design main removed because it corrupted the author's
  file on save. It now asserts what the pipeline actually owes: a dangerous
  scheme survives UNCHANGED (no rewrite, and no laundering into something
  that reads as safe) and is still classified unsafe, so the render and
  activation sinks refuse it. End-to-end containment is proven in
  services/navigation/linkSecurity.test.ts.

Reverted from main
- `diffSharedAttributes` widened `diff()` to compare shared attributes across
  a type divergence. Every ledger in this tier is measured against the
  narrower contract, so widening it here would re-author 20 commits of
  ledgers blind. The known gap it addressed is documented at the call site
  for the ledger layer to close.

Fidelity fix found by this branch's typed-input matrix
- the leading-`---` guard was OVER-BROAD: it rewrote any document-leading
  thematic break to `***`, so typing `---` in an empty document came back as
  `***`. It now VERIFIES — reparse and swap only when the break would really
  be read as frontmatter. A lone `---` is kept as written.

Also
- referenceType was a bare `string` cast, which this branch's own
  `satisfies LinkReference` correctly rejected; narrowed to the mdast union
- tsconfig excludes `__tests__` directories: test-support modules there use
  node APIs the browser build has no types for (3 instances, one class)
- runCases.ts declared as a knip entry — it is spawned as a child process,
  so no import makes it reachable
- 8 unused exports/types made module-private; knip types baseline 61 -> 60
- pmInlineConverters split at 336 lines: mark-edge-whitespace handling moved
  to markEdgeWhitespace.ts
- GHA corpus count 33 -> 34 and soak.yml recorded byte-identical
