+++ src/lib/ghaWorkflow/save/__tests__/corpusRoundtrip.test.ts
@@ -33,7 +33,7 @@ import { walkWorkflows, commentSet, anchorUsage } from "@/test/ghaCorpusHelpers"
 const CORPUS_ROOTS = ["src/test/fixtures/gha-workflows", ".github/workflows"];
 
 /** Pinned so a file vanishing from either root fails rather than shrinking the sweep. */
-const EXPECTED_FILE_COUNT = 33;
+const EXPECTED_FILE_COUNT = 34;
 
 /**
  * Workflows whose no-op round trip is BYTE-identical today. Two-way
@@ -42,6 +42,7 @@ const EXPECTED_FILE_COUNT = 33;
 const BYTE_IDENTICAL: ReadonlySet<string> = new Set([
   ".github/workflows/ci.yml",
   ".github/workflows/claude-cost-report.yml",
+  ".github/workflows/soak.yml",
   ".github/workflows/release.yml",
   ".github/workflows/rust-coverage.yml",
   ".github/workflows/star-history.yml",
